- #7 | Add Ontology Graph
Routing
- Added nested routes for tags of a given user and updated associated form
Refactoring
- Broke the stack of notes with chat input into separate file
- stack now appears on all pages independently of library

Layout / Navigation
- Dropdown menu added
- link to tags
- link to standalone stack of notes
- link to full library

Assets/JS:
- Split javascript into chunks using webpack
- Importing packs in body of page instead of head to ensure multi-select works (not why this is necessary?)

CSS:
- Reduce filter size
- Ensure colours applied to links on buttons
- Max height is set on the list itself instead of the library or the outer container
