+++ app/assets/stylesheets/application.css.erb
@@ -136,11 +136,17 @@ pre {
   border-left: 2px solid #69c;
 }
 
-a {
+a, a:visited {
   color: #e81c4f;
 }
-
-
+#library #entries {
+  overflow-y: scroll;
+  height: 75vh;
+}
+.message form {
+  padding-left: 1em;
+  padding-right: 1em;
+}
 <%# ------------------ %>
 
 html body .bp4-popover2 a, html body .bp4-popover2 li, html body .bp4-popover2 p, .bp4-popover2 a { color: unset; }
@@ -150,7 +156,11 @@ html body div div.bp4-active {
   background: transparent !important
 }
 
-body div#root div filter-input input { width: 40%; margin: 2%; }
+body div#root div filter-input input { 
+  width: 40%; 
+  margin: 2%;
+  height: fit-content;
+}
 .stack-biblio  { 
                 display : flex;
                 justify-content: space-around; 
@@ -189,6 +199,11 @@ body div#root div filter-input input { width: 40%; margin: 2%; }
 filter-input input { width: 100% }
 
 
+.thredded--currently-online {
+  right: unset;
+  left: 0;
+}
+
 @media only screen and (min-width: 992px) {
     details.chat-toggle {
       right: 5em;
@@ -201,11 +216,11 @@ filter-input input { width: 100% }
 /*-------------------- Chat --------------------*/
 details.chat-toggle {
   position: absolute;
-  bottom: 0;
+  bottom: 1em;
   right: 5em;
 }
 details.chat-toggle > summary {
-  background: rgba(255, 203, 165, 0.64);
+  background: rgba(255, 203, 165, 0.9);
   display: flex;
   border-radius: 10%;
   align-items: center;
@@ -397,7 +412,11 @@ details.chat-toggle > summary {
 .message-box .message-input {
   justify-content: space-between;
 }
-.tagsHolder, .message-input {
+.tagsHolder {
+  margin-right: 2%;
+  margin-left: 2%;
+} 
+.message-input {
   margin: 2%;
 }
 .message-box input:focus:-webkit-placeholder {
