+++ app/controllers/tags_controller.rb
@@ -63,7 +63,7 @@ class TagsController < ApplicationController
 
   # Use callbacks to share common setup or constraints between actions.
   def set_tag
-    @tag  = Tag.find(params[:id])
+    @tag = Tag.find(params[:id])
   end
 
   # Only allow a list of trusted parameters through.
