+++ app/views/layouts/partials/_navbar.html.erb
@@ -1,8 +1,15 @@
     
 <div class="flex justify-apart">
-  <span>
+  <span class="flex">
     <a class="button-like" href="/">InvolvedTowns Forums</a>
-    <a class="button-like" href="/library">My StackBiblio Library</a>
+    <details>
+    <summary class="button-like">StackBiblio</summary>
+    <ul>
+      <li><%= link_to('My Note Library', '/library') %></li>
+      <li><%= link_to('My Note Stack', '/stack') %></li>
+      <li><%= link_to('My Tags', Rails.application.routes.url_helpers.user_tags_path(user_id: current_user&.id)) %></li>
+    </ul>
+    </details>
   </span>
   <span>  
     <a class="button-like" href="/about">About</a>
