+++ app/views/layouts/thredded/application.html.erb
@@ -4,6 +4,11 @@
     <title>Forum <%= yield :thredded_page_title %></title>
     <meta content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0' name='viewport' />
     <%= stylesheet_link_tag 'thredded', 'data-turbolinks-track': 'reload' %>
+    <!-- Style dependencies -->
+    <link href="https://unpkg.com/normalize.css@^8.0.1" rel="stylesheet" />
+    <!-- Blueprint stylesheets -->
+    <link href="https://unpkg.com/@blueprintjs/icons@^4.0.0/lib/css/blueprint-icons.css" rel="stylesheet" />
+    <link href="https://unpkg.com/@blueprintjs/core@^4.0.0/lib/css/blueprint.css" rel="stylesheet" />
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= csrf_meta_tag %>
     <%= csp_meta_tag %>
@@ -18,5 +23,7 @@
   <body>
     <%= render('layouts/partials/navbar') %> 
     <%= yield %>
+    <%= javascript_packs_with_chunks_tag('chat') %>     <%# This has to be in the body instead of the head? %>
+    <div id="react-chat-container"></div>
   </body>
 </html>
