+++ app/views/stackbiblio/index.html.erb
@@ -1,2 +1,2 @@
-<%= javascript_pack_tag('stackbiblio') %>
-<div id="root"></div>
+<%= javascript_packs_with_chunks_tag('library') %>
+<div id="library"></div>
