+++ app/views/tags/_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(model: tag, local: true) do |form| %>
+<%= form_with(model: [current_user, tag], local: true) do |form| %>
   <% if tag.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(tag.errors.count, "error") %> prohibited this tag from being saved:</h2>
