+++ app/views/tags/index.html.erb
@@ -28,4 +28,4 @@
 
 <br>
 
-<%= link_to 'New Tag', new_tag_path %>
+<%= link_to 'New Tag', new_user_tag_path %>
