+++ app/views/tags/new.html.erb
@@ -2,4 +2,4 @@
 
 <%= render 'form', tag: @tag %>
 
-<%= link_to 'Back', tags_path %>
+<%= link_to 'Back', user_tags_path %>
