+++ config/routes.rb
@@ -1,12 +1,15 @@
 # frozen_string_literal: true
 
 Rails.application.routes.draw do
-  resources :tags
   devise_for :users
   mount Thredded::Engine => '/home'
 
-  resources :users
+  resources :users do
+    resources :tags
+  end
+  
   get '/library', to: 'stackbiblio#index'
+  get '/stack', to: 'stackbiblio#stack'
   get '/me', to: 'users#me', as: :me
   get '/about', to: 'home#index'
   root to: redirect('/home')
